requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
# cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-braces==1.15.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-jazzmin==2.4.8
django-oauth-toolkit==1.5.0
django-oauth2-provider==0.2.6.1
django-oidc-provider==0.7.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
drf-yasg==1.21.3
future==0.18.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pycparser==2.21
pycryptodomex==3.15.0
pyjwkest==1.4.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.14.0
